codecheck/database/requirements.txt
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
geoalchemy2==0.14.2
python-dotenv==1.0.0